\name{dat}
\alias{dat}
\docType{data}
\title{ German investment income consumption in log difference}
\description{
Lutkepohl's data
}
\usage{data(dat)}


\references{
Lutkepohl, H. 2005, New Introduction to Multiple Time Series Analysis, Springer
}
\examples{
data(dat)
}
\keyword{ts}
