\name{VAR.etp-package}
\alias{VAR.etp-package}
\alias{VAR.etp}
\docType{package}
\title{VAR modelling: estimation, testing, and prediction
}
\description{ Estimation, Hypothesis Testing, Prediction in Stationary Vector Autoregressive Models
}
\details{
\tabular{ll}{
Package: \tab VAR.etp\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-14\cr
License: \tab GPL-2 \cr
}

The data set dat.rda is from Lutkepohl's book
It is German Macrodata in log difference

Bootstrap bias-correction and prediction intervals are also included.

}
\author{
Jae H. Kim

Maintainer:  Jae H. Kim
}

\keyword{ package }

