\name{data1}
\alias{data1}
\docType{data}
\title{ stock return data used in Kim (2014)}
\description{
stock return data used in Kim (2014)
}
\usage{data(data1)}


\references{
Kim, J.H. 2014, Testing for parameter restrictions in a stationary VAR model: a bootstrap alternative. Economic Modelling, 41, 267-273.
}
\examples{
data(data1)
}
\keyword{ts}
