\name{VAR.etp-package}
\alias{VAR.etp-package}
\alias{VAR.etp}
\docType{package}
\title{VAR Modelling: Estimation, Testing, and Prediction
}
\description{ Estimation, Hypothesis Testing, Prediction in Stationary Vector Autoregressive Models
}
\details{
\tabular{ll}{
Package: \tab VAR.etp\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2023-08-31\cr
License: \tab GPL-2 \cr
}

The data set dat.rda is from Lutkepohl's book.

It is German Macrodata in log difference.

Bootstrap bias-correction and prediction intervals are also included.

Estimation and Forecasting based on Predictive Regression is also included.
}
\author{
Jae H. Kim

Maintainer:  Jae H. Kim
}

\keyword{ package }

